[feat/#26] Fix: cycle reference error
